[refactor] Complete Phase 1 - remove old multimodal_trainer.py
Deleted multimodal_trainer.py (2,927 lines)
Replaced with 6 modular files (2,549 lines)
Net reduction: 378 lines (13%)

Updated 3 import locations to use new modular architecture.
Import test passes. Phase 1 migration complete.
